# front/public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AdoptMe API Documentation",
    "description": "Especificación interactiva de los endpoints de la plataforma de adopción de mascotas AdoptMe. Incluye gestión de usuarios, roles, catálogo de mascotas y solicitudes de adopción.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de desarrollo local"
    }
  ],
  "paths": {
...
    "/api/adopciones": {
      "get": {
        "summary": "Listar solicitudes de adopción",
        "description": "Obtiene la lista de postulaciones. Permite filtrar las postulaciones de un usuario específico.",
        "parameters": [
          {
            "name": "usuario_id",
            "in": "query",
            "description": "ID del usuario cliente para filtrar sus propias solicitudes",
            "required": false,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de solicitudes obtenida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Adopcion"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Enviar una solicitud de adopción",
        "description": "Crea una postulación de un cliente para adoptar a una mascota disponible.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mascota_id": { "type": "string", "format": "uuid", "example": "5411a704-08de-433e-b8de-94dc281e0574" },
                  "usuario_id": { "type": "string", "format": "uuid", "example": "dd906b48-0afe-4c59-b9dd-300ffcaa169d" },
                  "mensaje": { "type": "string", "example": "Prometo cuidarla muy bien y darle mucho cariño." }
                },
                "required": ["mascota_id", "usuario_id"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Solicitud creada con éxito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Adopcion"
                }
              }
            }
          },
          "400": {
            "description": "mascota_id y usuario_id son requeridos"
          },
          "409": {
            "description": "Ya existe una postulación activa (pendiente o aprobada) para esta mascota"
          }
        }
      }
    },
    "/api/adopciones/{id}": {
      "get": {
        "summary": "Obtener solicitud de adopción por ID",
        "description": "Obtiene la información de una solicitud de adopción específica.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID único de la solicitud de adopción",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Solicitud obtenida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Adopcion"
                }
              }
            }
          },
          "404": {
            "description": "Solicitud no encontrada"
          }
        }
      },
      "put": {
        "summary": "Actualizar estado de solicitud (Aprobar/Rechazar)",
        "description": "Permite a la empresa actualizar el estado de una solicitud recibida.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID de la solicitud a actualizar",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "estado": { "type": "string", "enum": ["pendiente", "aprobada", "rechazada"], "example": "aprobada" }
                },
                "required": ["estado"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estado actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Adopcion"
                }
              }
            }
          },
          "400": {
            "description": "Estado inválido o no suministrado"
          },
          "404": {
            "description": "Solicitud de adopción no encontrada"
          }
        }
      },
      "delete": {
        "summary": "Eliminar solicitud",
        "description": "Elimina una postulación de adopción.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID de la solicitud",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "204": {
            "description": "Eliminado con éxito (Sin contenido)"
          },
          "404": {
            "description": "Solicitud no encontrada"
          }
        }
      }
    },
    "/api/adopciones/empresa": {
      "get": {
        "summary": "Listar solicitudes recibidas por empresa",
        "description": "Retorna todas las postulaciones asociadas a las mascotas publicadas por una empresa específica.",
        "parameters": [
          {
            "name": "empresa_id",
            "in": "query",
            "description": "ID único de la empresa/refugio",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de solicitudes de la empresa obtenida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Adopcion"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Adopcion": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "mascota_id": { "type": "string", "format": "uuid" },
          "usuario_id": { "type": "string", "format": "uuid" },
          "estado": { "type": "string", "example": "pendiente" },
          "mensaje": { "type": "string", "example": "Mensaje de postulación" }
        }
      }
    }
  }
}